+++ packages/terra-functional-testing/src/terra-cli/wdio/test-runner.js
@@ -104,7 +104,7 @@ class TestRunner {
     } = options;
 
     // Clean only the non reference screenshots.
-    cleanScreenshots();
+    cleanScreenshots(options.useRemoteReferenceScreenshots);
 
     await TestRunner.configureScreenshots(options);