+++ packages/terra-functional-testing/CHANGELOG.md
@@ -5,6 +5,7 @@
 * Added
   * Added authentication for accessing screenshots from the remote site.
   * Added `useRemoteReferenceScreenshots` cli option for downloading reference screenshots from a remote site for screenshot comparisons.
+  * Delete the `reference` screenshot directory when `useRemoteReferenceScreenshots` is true.
 
 * Fixed
   * Fixed deleting the `diff`, `error`, and `latest` screenshot directories at the beginning of each test run.